requirements.txt
requests
pdf2image
Pillow
pyzbar
streamlit